requirements.txt
Babel==2.6.0
dj-database-url==0.5.0
Django==2.2.5
django-classy-tags==0.9.0
django-filer==1.6.0
django-formtools==2.1
django-garpix-helpers==0.1.3
django-garpix-slider==0.1.14
django-hvad==2.0.0
django-js-asset==1.2.2
django-libs==2.0.2
django-mptt==0.10.0
django-phonenumber-field==2.2.0
django-polymorphic==2.0.3
django-sekizai==1.0.0
django-treebeard==4.3
django-uuslug==1.2.0
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-column==1.9.0
djangocms-installer==1.1.1
djangocms-snippet==2.1.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.7.0
easy-thumbnails==2.6
gunicorn==19.9.0
html5lib==1.0.1
numpy==1.17.4
phonenumbers==8.10.6
Pillow==5.4.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-magic==0.4.15
python-slugify==4.0.0
pytz==2018.7
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
tzlocal==2.0.0
Unidecode==1.0.23
webencodings==0.5.1
